# api/openapi.yaml
openapi: 3.1.0
info:
  title: memplua Local API
  version: 1.0.0
  description: Loopback-only REST/SSE API. Desktop clients send X-Memplua-Client=desktop for safe localized error codes; the developer UI contracts remain available. Every API endpoint except /healthz requires the local bearer token.
servers:
  - url: http://127.0.0.1:7331
security:
  - bearerAuth: []
paths:
...
  /api/v1/sources/{id}/start:
    post:
      operationId: startSource
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "202":
          description: Source session accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceSession" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/source-sessions/{id}/pause:
    post:
      operationId: pauseSourceSession
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204": { description: Session paused. }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/source-sessions/{id}/resume:
    post:
      operationId: resumeSourceSession
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "202":
          description: Existing session resumed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceSession" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/source-sessions/{id}/stop:
    post:
      operationId: stopSourceSession
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204": { description: Session stopped. }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/ingest/text:
    post:
      operationId: submitManualText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 250000 }
      responses:
        "202":
          description: Text was persisted as a chunk and durable grouping was queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TextSubmission" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/source-sessions/{id}/finalize:
    post:
      operationId: finalizeSourceSession
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Idempotently close the collecting batch; capture continues.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/conspects:
    get:
      operationId: listReviewConspects
      parameters:
        - $ref: "#/components/parameters/DesktopView"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Oldest first conspects. view=desktop returns compact summaries; omitted view preserves the developer contract.
          content:
            application/json:
              schema:
                oneOf:
                  - { type: array, items: { $ref: "#/components/schemas/ReviewConspect" } }
                  - { type: array, items: { $ref: "#/components/schemas/DesktopConspectSummary" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/conspects/{id}:
    get:
      operationId: getReviewConspect
      parameters:
        - $ref: "#/components/parameters/DesktopView"
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Structured candidates, matches, decisions and taxonomy.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ReviewConspect" }
                  - { $ref: "#/components/schemas/DesktopConspectDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/conspects/{id}/source-text:
    get:
      operationId: getConspectSourceText
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Full immutable focus text in sequence order. No technical IDs, raw JSON, model responses or context tail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConspectSourceText" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/review/conspects/{id}/chunks:
    get:
      operationId: getConspectEvidence
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Original evidence chunks.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Chunk" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/items/{id}:
    put:
      operationId: saveReviewDecision
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewDecision" }
      responses:
        "204":
          description: Save a decision without changing the canonical graph.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/items/{id}/split:
    post:
      operationId: splitReviewVariants
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SplitReviewItem" }
      responses:
        "204":
          description: Move a proper subset of incoming variants into another review item.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/taxonomy/{id}:
    put:
      operationId: resolveReviewTaxonomy
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaxonomyDecision" }
      responses:
        "204":
          description: Resolve a unique unknown label without changing taxonomy.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/review/conspects/{id}/apply:
    post:
      operationId: applyConspect
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Atomically apply all resolved items, taxonomy, links, provenance, one graph revision and one export job. Repeated application is a no-op; a version conflict rolls back and queues rematching.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/graph/search:
    get:
      operationId: searchCanonicalGraph
      parameters:
        - name: kind
          in: query
          required: true
          schema: { enum: [term, thought] }
        - name: q
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Search the complete canonical graph by normalized title or body.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ReviewMatch" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/developer/analysis-batches:
    get:
      operationId: listAnalysisBatches
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Developer diagnostics including raw and normalized responses.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AnalysisBatch" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/developer/analysis-batches/{id}:
    get:
      operationId: getAnalysisBatch
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Developer diagnostics and original focus/context fragments.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalysisBatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/jobs:
    get:
      operationId: listJobs
      parameters:
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema: { enum: [queued, running, retry, done, failed] }
      responses:
        "200":
          description: Durable processing jobs, optionally filtered by status.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
...
  /api/v1/jobs/{id}/retry:
    post:
      operationId: retryFailedJob
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "202": { description: Failed job returned to the queue. }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/notifications/{id}/read:
    post:
      operationId: markNotificationRead
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204": { description: Notification marked as read. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict or validation failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    TextSubmission:
      type: object
      required: [source_session_id, chunk_id, created_at]
      properties:
        source_session_id: { type: string }
        chunk_id: { type: string }
        created_at: { type: string, format: date-time }
...
    SourceSession:
      type: object
      required: [id, source_id, kind, state, started_at]
      properties:
        id: { type: string }
        source_id: { type: string }
        kind: { type: string }
        state: { enum: [starting, running, paused, stopped, failed] }
        started_at: { type: string, format: date-time }
        stopped_at: { type: string, format: date-time }
        last_error: { type: string }
...